feat: move application service interfaces to to dependent folder
