+++ src/application/interfaces/interpolatorInterface.ts
